refactor: splitted into files the engine models, while keeping retro-compatibility with the previous path
