+++ src/engine/movie.ts
@@ -1,0 +1,5 @@
+import { PriceCode } from "./price-code";
+
+export class Movie {
+    constructor(public readonly title: string, public readonly priceCode: PriceCode) { }
+}
