+++ src/engine/price-code.ts
@@ -1,0 +1,3 @@
+export class PriceCode {
+    constructor(public readonly name: string) {}
+}
