+++ src/engine/rental.ts
@@ -1,0 +1,5 @@
+import { Movie } from "./movie";
+
+export class Rental {
+    constructor(public readonly movie: Movie, public readonly daysRented: number) { }
+}
